+++ src/Controller/Admin/BannersPositionsController.php
@@ -75,7 +75,6 @@ class BannersPositionsController extends AppController {
     /**
      * Edits banners position
      * @param string $id Banners Position ID
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function edit($id = NULL)  {
         $position = $this->BannersPositions->get($id);
@@ -96,7 +95,6 @@ class BannersPositionsController extends AppController {
     /**
      * Deletes banners position
      * @param string $id Banners Position ID
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function delete($id = NULL) {
         $this->request->allowMethod(['post', 'delete']);