+++ src/Controller/Admin/UsersController.php
@@ -92,7 +92,7 @@ class UsersController extends AppController {
 			->contain(['Groups' => ['fields' => ['label']]])
 			->select(['id', 'username', 'email', 'first_name', 'last_name', 'active', 'banned', 'post_count', 'created'])
 			->where(['Users.id' => $id])
-			->first()
+			->firstOrFail()
         );
     }
 
@@ -120,13 +120,12 @@ class UsersController extends AppController {
      * Edits user
      * @param string $id User ID
 	 * @uses MeCms\Controller\Component\AuthComponent::isFounder()
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function edit($id = NULL)  {
 		$user = $this->Users->find()
 			->select(['id', 'group_id', 'username', 'email', 'first_name', 'last_name', 'active'])
 			->where(['Users.id' => $id])
-			->first();
+			->firstOrFail();
 		
 		//Only the admin founder can edit others admin users
 		if($user->group_id === 1 && !$this->Auth->isFounder()) {
@@ -155,7 +154,6 @@ class UsersController extends AppController {
      * Deletes user
      * @param string $id User ID
 	 * @uses MeCms\Controller\Component\AuthComponent::isFounder()
-     * @throws \Cake\Network\Exception\NotFoundException
      */
     public function delete($id = NULL) {
         $this->request->allowMethod(['post', 'delete']);
@@ -163,7 +161,7 @@ class UsersController extends AppController {
 		$user = $this->Users->find()
 			->select(['id', 'group_id', 'post_count'])
 			->where(['Users.id' => $id])
-			->first();
+			->firstOrFail();
 		
 		//You cannot delete the admin founder
 		if($user->id === 1)
@@ -186,10 +184,10 @@ class UsersController extends AppController {
 	/**
 	 * Activates account
 	 * @param string $id User ID
-     * @throws \Cake\Network\Exception\NotFoundException
 	 */
 	public function activate($id) {
 		$user = $this->Users->get($id);
+		
 		$user->active = TRUE;
 		
 		if($this->Users->save($user))
@@ -208,15 +206,14 @@ class UsersController extends AppController {
 		$user = $this->Users->find()
 			->select(['id', 'email', 'first_name', 'last_name'])
 			->where(['id' => $this->Auth->user('id')])
-			->first();
+			->firstOrFail();
 		
         if($this->request->is(['patch', 'post', 'put'])) {
 			$user = $this->Users->patchEntity($user, $this->request->data);
 			
 			if($this->Users->save($user)) {
 				//Sends email
-				$this->getMailer('MeCms.User')
-					->send('change_password', [$user]);
+				$this->getMailer('MeCms.User')->send('change_password', [$user]);
 				
 				$this->Flash->success(__d('me_cms', 'The password has been edited'));
 				return $this->redirect(['_name' => 'dashboard']);