+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.1 branch
 ### 2.1.7
+* an exception is now properly thrown when a record is not found;
 * fixed bug. Now managers can access "system" menu.
 
 ### 2.1.6